Centered, 8 bit gradient-colored label with adjustable center offset
The program enables the alternate buffer, hides the cursor, disables the
terminal's echo, displays the number of rows and columns, the colors the
entire grid with alternating colors to ease the counting of the cells.
It then determines the propert left offset in order to print a label of
variable length (which is the first argument of the program or "Hello
World!" is there is none) in the center of the terminal, according to
the given horizontal alignment margin of error, if the label cannot be
perfectly centered. Also note that the label uses a gradient of 8 bit
color values from "black" to "white" (refer to ANSI specification for
this). It then determines the right row according to the specified
vertical offset margin of error, if the label cannot be perfectly
virticaly centered.
The program then waits for the enter key to be pressed and restores the
terminal's original state (re-enable echo, show normal buffer and
cursor).