Utf8 Support for Windows
Adds utf8 support for Windows using SetConsoleOutputCP(CP_UTF8) (code
page 65001). Note however that utf8 argument parsing in the commandline
is not possible if the commandline's codepage is not set to 65001, which
is impossible to check beforehand, before the program's execution, of
course. New terminal windows could be launched with utf8 support though.
This doesn't have anything to do with internal utf8 support, the porgram
will still work fine on Windows, that is after `SetConsoleOutputCP` is
called.