Add the (ro_)termsize struct and get_termsize, ref_termsize
Very simple in nature, stores the number of rows/columns and the size of
the terminal window in pixels.
On unix system, this is already what the winsize struct is for, so it's
a simple remapping. A tiny bit more work is needed to determin the pixel
size of the window on Windows.
the ro_- variant has constant fields, so that they cannot be modified
after initialization. It can only be used with `get_termsize`, as it
creates the struct on return. It has its advantages, and so does
`ref_termsize`.