Makefile can now updates archives incrementally instead of recreating them
Previously, the Makefile used to always overwrite archives (static libs)
by creating a new one with all the object files inside.
Now it updates the archive file (using ar) :
- If the archive does not exist, it is created.
- For all the objects declared as prequisite for the static lib target,
if one of them is not in the archive yet, it is inserted.
- If the archive member is outdated, it is replaced by the newly
compiled one.

Although you WILL never notice the difference (except in the cmdline as
it prints less text since there are fewer arguments passed to ar), it
still rocks.