Restrict the format of CC
CC now discards dashes '-' when prefixed and will only take the first
"word" into account, storing it into a "CCC" variable (Could stand for
Cut CC but idk)
A couple examples:
`> make CC=-gcc -O333` will assign "gcc" to CCC
`> make CC=blabla` will return an error when trying to compile if blabla
does not exist or is not a drop-in replacement for gcc (like clang or
zig's cc are)

Correction : CC should be whatever the user provided

I actually over-engineered this : some drop-in replacement for gcc like
`zig cc` (that I actually mentionned above) is, as you can tell, two
"words". Maybe some day there will be a compiler that does this with
dashes : `fancy_compiler --as-cc`. So I'll trust the developper for not
putting weird stuff in CC (seems reasonable enough)
AR is now also overrideable in the cmdline