Branchless off-center alignment
Using enum values which correspond to the correct offset when the
label's length parity and the terminal's differ allows for a branchless
addition.
The separation into two enums for the margin of error preference allows
for better type safety and warnings, although the syntax might be less
*sugary*.
Other definitions in `label.h` will become relevant in a few commits.